+++ VenteAuCess.html
@@ -3,8 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <script type="module" src="script.js"></script>
-        <script src="theme_changer.js"></script>
-        <link rel="stylesheet" href="styles.css" />
+        <link rel="stylesheet" href="assets/css/old.css" />
         <link
             rel="stylesheet"
             href="https://fonts.googleapis.com/css2?family=Roboto+Mono&display=swap"
@@ -33,6 +32,7 @@
                         target="_blank"
                         >Instagram : livres_cess_ouest</a
                     >
+                    <a href="index.html">test</a>
                 </span>
                 <span id="themechange">
                     <button onclick="changeTheme()">Theme</button>
@@ -47,10 +47,12 @@
             <div id="main">
                 <!-- ---------------------------------------------------------------------------------------------------- -->
                 <div class="content">
-                    <a href="page article/Anglais Le mot pour dire.html">
+                    <a
+                        href="page article/English/Le Mot Pour Dire/Anglais Le mot pour dire.html"
+                    >
                         <img
                             class="imgArticle"
-                            src="page article/images/Anglais Le mot pour dire 1.jpg"
+                            src="page article/English/Le Mot Pour Dire/images/Anglais Le mot pour dire 1.jpg"
                             alt="Anglais Le mot pour dire"
                         />
                         <div class="description">
@@ -64,7 +66,7 @@
                     <a href="page article/Atlas mondial suisse.html">
                         <img
                             class="imgArticle"
-                            src="page article/images/Atlas mondial suisse 1.jpg"
+                            src="page article/Other/Atlas Mondial Suisse/images/Atlas mondial suisse 1.jpg"
                             alt="Atlas mondial suisse"
                         />
                         <div class="description">
@@ -78,7 +80,7 @@
                     <a href="page article/CCCO Edition commerciale.html">
                         <img
                             class="imgArticle"
-                            src="page article/images/CCCO Edition commerciale 1.jpg"
+                            src="page article/Eco Droit/CCCO Edition Commerciale/images/CCCO Edition commerciale 1.jpg"
                             alt="CCCO Edition commerciale"
                         />
                         <div class="description">
@@ -94,7 +96,7 @@
                     >
                         <img
                             class="imgArticle"
-                            src="page article/images/Dictionnaire Anglais-Français Larousse 200'000 mots 1.jpg"
+                            src="page article/English/Larousse Anglais-Français 200 000 mots/images/Dictionnaire Anglais-Français Larousse 200'000 mots 1.jpg"
                             alt="Dictionnaire Anglais-Français Larousse 200'000 mots"
                         />
                         <div class="description">
@@ -113,7 +115,7 @@
                     >
                         <img
                             class="imgArticle"
-                            src="page article/images/Dictionnaire Français Le Robert Poche 75'000 mots 1.jpg"
+                            src="page article/Français/LeRobert Dictionnaire Français 75 000 mots/images/Dictionnaire Français Le Robert Poche 75'000 mots 1.jpg"
                             alt="Dictionnaire Français Le Robert Poche 75'000 mots"
                         />
                         <div class="description">
@@ -130,7 +132,7 @@
                     <a href="page article/Code des obligations suisses.html">
                         <img
                             class="imgArticle"
-                            src="page article/images/Code des obligations suisses 1.jpg"
+                            src="page article/Eco Droit/Code des Obligations Suisses/images/Code des obligations suisses 1.jpg"
                             alt="Code des obligations suisses"
                         />
                         <div class="description">
