+++ assets/css/old.css
@@ -97,18 +97,3 @@ button:hover {
 .imglogo {
     height: 100%;
 }
-/*---------------------------------------------*/
-/* New Part*/
-
- .item-list {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: 20px;
-  }
-  
-  .item {
-    border: 1px solid #ccc;
-    padding: 10px;
-    text-align: center;
-  }
-
