+++ page article/English/Le Mot Pour Dire/Anglais Le mot pour dire.html
@@ -4,7 +4,8 @@
         <meta charset="utf-8" />
         <title>Anglais Le mot pour dire</title>
         <link rel="icon" href="icone.jpg" />
-        <link rel="stylesheet" href="style.css" />
+        <link rel="stylesheet" href="page article/styles_articles_sample.css" />
+        <link rel="stylesheet" href="../../styles_articles_sample.css" />
         <script src="script.js"></script>
         <script>
             var article = "Anglais Le mot pour dire";
